open-unions: 3.2: polymorphic single-effects
in the process of implementing this ran into scala's limitations
regd type constructor inference. specifically, the Reader effect could
not be represented as a simple trait w/ kind-projector style holes. the
workaround is documented under sss.workarounds.TypeConstructorInference
and used in the definition of Reader effect in sss.effects.Poly