open-unions: 3.2: fix monad implicit for Eff
restructuring Eff in a manner similar to Reader fixes the issue with the
monad implicits not firing for Eff. it seems the underlying issue there
too was regd scala's unification engine having issues with type
constructors. this comes at the cost of slightly greater type
annotation. specifically, 'ask' requires the type to be provided. there
are, however, a couple of unresolved issues:

1] FIXME: having implicits for both functor and monad should be
   ok. currently Eff's functor implicit, when disabled, breaks things.

2] TODO: are the inner traits such as Eff[r]#`0` needed? if not, let's
   simplify